+++ cmd/cascade-registry/main.go
@@ -10,11 +10,11 @@ import (
 	"strconv"
 	"strings"
 
-	v2 "github.com/robinkb/cascade-registry/api/v2"
 	"github.com/robinkb/cascade-registry/cluster/raft"
 	"github.com/robinkb/cascade-registry/cluster/raft/qwal"
 	"github.com/robinkb/cascade-registry/process"
 	"github.com/robinkb/cascade-registry/registry"
+	v2 "github.com/robinkb/cascade-registry/registry/api/v2"
 	"github.com/robinkb/cascade-registry/server"
 	"github.com/robinkb/cascade-registry/store/boltdb"
 	"github.com/robinkb/cascade-registry/store/cluster"
