+++ registry/api/v2/handler_test.go
