+++ registry/api/v2/manifests_test.go
