+++ registry/api/v2/tags_test.go
