+++ registry/api/v2/upload_sessions_test.go
