+++ testing/conformance/pull_test.go
@@ -5,8 +5,8 @@ import (
 	"strconv"
 	"testing"
 
-	v2 "github.com/robinkb/cascade-registry/api/v2"
 	"github.com/robinkb/cascade-registry/registry"
+	v2 "github.com/robinkb/cascade-registry/registry/api/v2"
 	"github.com/robinkb/cascade-registry/store/inmemory"
 	. "github.com/robinkb/cascade-registry/testing"
 	testclient "github.com/robinkb/cascade-registry/testing/client"
