+++ testing/conformance/push_test.go
@@ -7,8 +7,8 @@ import (
 	"net/http"
 	"testing"
 
-	v2 "github.com/robinkb/cascade-registry/api/v2"
 	"github.com/robinkb/cascade-registry/registry"
+	v2 "github.com/robinkb/cascade-registry/registry/api/v2"
 	"github.com/robinkb/cascade-registry/store/inmemory"
 	. "github.com/robinkb/cascade-registry/testing"
 	testclient "github.com/robinkb/cascade-registry/testing/client"
